+++ bin/dart_language_server.dart
@@ -2,11 +2,19 @@ import 'dart:io';
 import 'dart:async';
 
 import 'package:dart_language_server/dart_language_server.dart';
+import 'package:dart_language_server/src/args.dart';
 
-Future main() async {
+Future main(List<String> args) async {
+  StartupArgs startupArgs;
+  try {
+    startupArgs = new StartupArgs(args);
+  } on UsageException catch (e) {
+    print(e.usage);
+    return;
+  }
   await runZoned(() async {
     try {
-      var shim = await startShimmedServer();
+      var shim = await startShimmedServer(startupArgs);
       await new StdIOLanguageServer.start(shim).onDone;
     } catch (e, st) {
       await new File('/tmp/lsp-error.log').writeAsString('Caught $e\n$st');