+++ lib/dart_language_server.dart
@@ -1,3 +1,3 @@
 export 'src/protocol/language_server/server.dart';
-export 'src/shim.dart';
+export 'src/shim.dart' show startShimmedServer;
 export 'src/logging/logs.dart' show closeLogs;