+++ lib/src/shim.dart
@@ -6,6 +6,7 @@ import 'package:analysis_server_lib/analysis_server_lib.dart'
     hide Position, Location;
 
 import 'apply_change.dart';
+import 'args.dart';
 import 'capabilities.dart';
 import 'logging/logs.dart';
 import 'position_convert.dart';
@@ -13,23 +14,25 @@ import 'protocol/language_server/interface.dart';
 import 'protocol/language_server/messages.dart';
 import 'utils/file_cache.dart';
 
-Future<LanguageServer> startShimmedServer() async {
+Future<LanguageServer> startShimmedServer(StartupArgs args) async {
   var client = await AnalysisServer.create(
       onRead: (m) => analyzerSink.add('OUT: $m\n'),
-      onWrite: (m) => analyzerSink.add('IN: $m\n'));
+      onWrite: (m) => analyzerSink.add('IN: $m\n'),
+      serverArgs: args.analysisServerArgs);
   await client.server.onConnected.first;
-  return new AnalysisServerAdapter(client);
+  return new AnalysisServerAdapter(client, args);
 }
 
 /// Wraps an [AnalysisServer] and exposes it as a [LanguageServer].
 class AnalysisServerAdapter implements LanguageServer {
   final AnalysisServer _server;
+  final StartupArgs _args;
 
   final _files = new FileCache();
 
   final _fileVersions = <String, int>{};
 
-  AnalysisServerAdapter(this._server) {
+  AnalysisServerAdapter(this._server, this._args) {
     _listeners();
   }
 
@@ -41,8 +44,8 @@ class AnalysisServerAdapter implements LanguageServer {
       ClientCapabilities clientCapabilities, String trace) async {
     final directory = _filePath(rootUri);
     _openDirectories.add(directory);
-    var clientName = '${p.basename(directory)}-$clientPid';
-    startLogging(clientName, trace);
+    final clientName = '${p.basename(directory)}-$clientPid';
+    startLogging(clientName, _args.forceTraceLevel ?? trace);
     await _server.analysis
         .setAnalysisRoots(_openDirectories.toList(), const []);
     return serverCapabilities;